Fix bug where new races with new teams would reset team counts
This would result in the config being created correctly but the new
round summary showing no teams for any division.

Adding a new team changes the league config which resulted in the reset
because it was wrapped in createEffect so this was removed.

Should maybe make a change to wrapagain but only trigger on a material
change or to preserve existing set teams if already set.