Remove circular import to fix tests, other minor enhancements
Don't need to "refetch" config when progressing or completing a stage,
just use the config on the round object.
Correct position strings when number ends in 11, 12 or 13.
createRound now accepts the "registry" of race configuration to draw
from which not only facilitates the removal of the circular dep but also
aligns with future state of supporting custom config groups.